Add an index to contexts and refactor corresponding imports
