+++ apps/user-portal/src/app.tsx
@@ -23,7 +23,7 @@ import { Redirect, Route, Router, Switch } from "react-router-dom";
 // import { Dimmer, Loader } from "semantic-ui-react";
 import { ProtectedRoute } from "./components";
 import { routes } from "./configs";
-import { AuthContext } from "./contexts/auth";
+import { AuthContext } from "./contexts";
 import { history, i18n } from "./helpers";
 import { store } from "./store";
